+++ tests/unit/test_iris_helpers.py
@@ -18,7 +18,7 @@ from iris.coords import (
     DimCoord,
 )
 from iris.cube import Cube, CubeList
-from iris.exceptions import CoordinateMultiDimError
+from iris.exceptions import CoordinateMultiDimError, UnitConversionError
 
 from esmvalcore.iris_helpers import (
     add_leading_dim_to_cube,
@@ -28,6 +28,7 @@ from esmvalcore.iris_helpers import (
     has_unstructured_grid,
     merge_cube_attributes,
     rechunk_cube,
+    safe_convert_units,
 )
 
 
@@ -571,3 +572,59 @@ def test_has_unstructured_grid_true(lat_coord_1d, lon_coord_1d):
         aux_coords_and_dims=[(lat_coord_1d, 0), (lon_coord_1d, 0)],
     )
     assert has_unstructured_grid(cube) is True
+
+
+@pytest.mark.parametrize(
+    "old_units,new_units,old_standard_name,new_standard_name,err_msg",
+    [
+        ("m", "km", "altitude", "altitude", None),
+        ("Pa", "hPa", "air_pressure", "air_pressure", None),
+        (
+            "m",
+            "DU",
+            "equivalent_thickness_at_stp_of_atmosphere_ozone_content",
+            "equivalent_thickness_at_stp_of_atmosphere_ozone_content",
+            None,
+        ),
+        (
+            "m",
+            "s",
+            "altitude",
+            ValueError,
+            r"Unable to convert from 'Unit\('m'\)' to 'Unit\('s'\)'",
+        ),
+        (
+            "unknown",
+            "s",
+            "air_temperature",
+            UnitConversionError,
+            r"Cannot convert from unknown units",
+        ),
+        (
+            "kg m-2 s-1",
+            "mm day-1",
+            "precipitation_flux",
+            ValueError,
+            r"Cannot safely convert units from 'kg m-2 s-1' to 'mm day-1'; "
+            r"standard_name changed from 'precipitation_flux' to "
+            r"'lwe_precipitation_rate'",
+        ),
+    ],
+)
+def test_safe_convert_units(
+    old_units, new_units, old_standard_name, new_standard_name, err_msg
+):
+    """Test ``esmvalcore.preprocessor._units.safe_convert_units``."""
+    cube = Cube(0, standard_name=old_standard_name, units=old_units)
+
+    # Exceptions
+    if isinstance(new_standard_name, type):
+        with pytest.raises(new_standard_name, match=err_msg):
+            safe_convert_units(cube, new_units)
+        return
+
+    # Regular test cases
+    new_cube = safe_convert_units(cube, new_units)
+    assert new_cube is cube
+    assert new_cube.standard_name == new_standard_name
+    assert new_cube.units == new_units
